% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{agreement}
\alias{agreement}
\title{Calculate percent agreement.}
\usage{
agreement(thedata)
}
\arguments{
\item{thedata}{rating matrix.}
}
\value{
percent rater agreement.
}
\description{
Calculate percent agreement.
}
