% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.IRRsim.R
\name{as.data.frame.IRRsim}
\alias{as.data.frame.IRRsim}
\title{Convert IRR simulation run to a data frame.}
\usage{
\method{as.data.frame}{IRRsim}(x, ...)
}
\arguments{
\item{x}{the result of \code{\link{simulateIRR}}.}

\item{...}{currently unused.}
}
\value{
a data.frame with the following columns:
\describe{
\item{nLevels}{}
\item{nEvents}{}
\item{k}{the number of raters used in the simulation.}
\item{k_per_event}{number of raters per scoring event}
\item{simAgreement}{the calculated percent agreement from the sample.}
\item{agreement}{the specified percent agreement used for drawing the random sample.}
\item{skewness}{skewness of all responses.}
\item{kurtosis}{Kurtosis for all responses.}
\item{MaxResponseDiff}{the difference between the most and least freqeuent responses.}
\item{ICC1}{ICC1 as described in Shrout and Fleiss (1979)}
\item{ICC2}{ICC2 as described in Shrout and Fleiss (1979)}
\item{ICC3}{ICC3 as described in Shrout and Fleiss (1979)}
\item{ICC1k}{ICC1k as described in Shrout and Fleiss (1979)}
\item{ICC2k}{ICC2k as described in Shrout and Fleiss (1979)}
\item{ICC3k}{ICC3k as described in Shrout and Fleiss (1979)}
\item{Fleiss_Kappa}{Fleiss' Kappa for m raters as described in Fleiss (1971).}
\item{Cohen_Kappa}{Cohen's Kappa as calculated in psych::cohen.kappa. Note that this
      calculated for all datasets even though it is only appropriate for two raters.}
}
}
\description{
Converts the results of [IRRsim::simulateIRR()] to a `data.frame`.
}
