% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRRsim-package.R
\docType{data}
\name{IRRguidelines}
\alias{IRRguidelines}
\title{Guidelines for interpreting inter-rater reliability statistics.}
\format{
A List object with numeric vectors containing the thresholds for
        interpreting IRR statistics. The names of the vectors correspond to
        the labels use by the authors. The name in the List object corresponds
        the first author's name. See vignette for more information.
}
\usage{
data(IRRguidelines)
}
\description{
Guidelines for interpreting inter-rater reliability statistics.
}
\examples{
data(IRRguidelines)
names(IRRguidelines) # Names of the guidelines
IRRguidelines[['Cicchetti']] # Cicchetti's (1994) guidelines
}
\keyword{ICC}
\keyword{IRR}
\keyword{datasets}
