% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRRsim-package.R
\docType{package}
\name{IRRsim-package}
\alias{IRRsim-package}
\title{Simulate Inter-Rater Reliability Statistics}
\description{
Simulate Inter-Rater Reliability Statistics
}
\author{
\email{jason@bryer.org}
}
\keyword{IRR}
\keyword{interrater}
\keyword{package}
\keyword{reliability}
\keyword{simulation}
