% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRRsim_demo.R
\name{IRRsim_demo}
\alias{IRRsim_demo}
\title{Run the Inter-Rater Reliability shiny application.}
\usage{
IRRsim_demo(clean = FALSE)
}
\arguments{
\item{clean}{if `TRUE` the cache file will be deleted when the application finishes.}
}
\description{
Run the Inter-Rater Reliability shiny application.
}
