% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRRsim-package.R
\docType{data}
\name{IRRsimData}
\alias{IRRsimData}
\title{Simulated data used to create tables of expected IRR metrics}
\format{A data.frame}
\usage{
data(IRRsimData)
}
\description{
This data frame is the results of simulating data sets with following parameters:
* Between 2 and 5 scoring levels
* Between 2 adn 12 raters
* Four response distributions including: uniform, lightly skewed, moderately
  skewed, and highly skewed.
}
\keyword{datasets}
