% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappan.fleiss2.R
\name{kappam.fleiss2}
\alias{kappam.fleiss2}
\title{Fleiss' Kappa for m raters}
\usage{
kappam.fleiss2(ratings, exact = FALSE, detail = FALSE)
}
\arguments{
\item{ratings}{n*m matrix or dataframe, n subjecs of m raters.}

\item{exact}{a logical indicating whether the exact Kappa (Conger, 1980) or the Kappa described
by Fleiss (1971) should be computed.}

\item{detail}{a logical indicating whether category-wise Kappas should be computed}
}
\description{
This function fixes an issue in the \code{kappam.fleiss} function in the \code{irr} package.
Specifically, the original function removes all missing values. However, many rating designs
do not have all raters score all essays. This function will still calculate Fleiss' Kappa for
those situations.
}
\details{
See \code{irr::kappam.fleis} for more information.
}
