% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.IRRsim.R
\name{plot.IRRsim}
\alias{plot.IRRsim}
\title{Plot the results of an IRR simulation.}
\usage{
\method{plot}{IRRsim}(x, method = "loess", stat, point.alpha = 0.3, title = TRUE, ...)
}
\arguments{
\item{x}{the result of \code{\link{simulateIRR}}.}

\item{method}{the method used to find predicted values. Possible values are
loess, linear, and quadratic.}

\item{stat}{the IRR statistic to return summary for, or "all" for all metrics.}

\item{point.alpha}{the alpha used for points.}

\item{title}{the title of the plot}

\item{...}{currently unused.}
}
\description{
Returns a \code{ggplot2} expression.
}
