% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lrdf.R
\name{plot.lrdf}
\alias{plot.lrdf}
\title{Logistic Regression Plot}
\usage{
\method{plot}{lrdf}(
  x,
  colors = c("#1b9e77", "#7570b3", "#d95f02"),
  rel_heights = c(0.4, 0.6),
  title,
  ...
)
}
\arguments{
\item{x}{the results of \code{\link{simulatePredictionData}}.}

\item{colors}{the color scheme used for the three groups.}

\item{rel_heights}{the relative heights of histogram to scatter plot.}

\item{title}{the title of plot.}

\item{...}{other parameters passed to \code{cowplot::plot_grid}.}
}
\value{
a ggplot2 plot of the logistic regression results.
}
\description{
Logistic Regression Plot
}
