% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.IRRsim.R
\name{print.IRRsim}
\alias{print.IRRsim}
\title{Prints the results from `simulateIRR.`}
\usage{
\method{print}{IRRsim}(x, ...)
}
\arguments{
\item{x}{the results of IRRsim::simulateIRR().}

\item{...}{currently unused.}
}
\description{
Prints the results from `simulateIRR.`
}
