% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePredictionData.R
\name{simulatePredictionData}
\alias{simulatePredictionData}
\title{Simulate data to test logistic regression.}
\usage{
simulatePredictionData(mean1 = -0.1, n1 = 100, mean2 = -mean1,
  n2 = n1, mean_unknown = (mean1 + mean2)/2, n_unknown = round(n1/2),
  ratio_unknown = 0.5, sd = 0.2, group_labels = c("A", "B", "NA"))
}
\arguments{
\item{mean1}{the mean of group 1.}

\item{n1}{the size of group 1.}

\item{mean2}{the mean of group 2.}

\item{n2}{the size of group 2.}

\item{mean_unknown}{the mean of the group with a random outcome.}

\item{n_unknown}{the size of the group with a random outcome.}

\item{ratio_unknown}{the probability of outcome 1 occurring in the random group.}

\item{sd}{the standard deviation for each group (i.e. parameter passed to
\code{rnorm}). Variance is assumed to be equal across all groups.}

\item{group_labels}{the labels for the groups in the \code{data.frame}.}
}
\value{
a \code{data.frame} with \code{n1 + n2 + n_unknown} rows and 3
       variables (\code{Group}, \code{X}, and \code{Y}).
}
\description{
Simulate data to test logistic regression.
}
