% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePredictionModels.R
\name{simulatePredictionModels}
\alias{simulatePredictionModels}
\title{Simulation for estimating logistic regression with uncertain responses.}
\usage{
simulatePredictionModels(nSamples = 500, ...)
}
\arguments{
\item{nSamples}{number of random datasets to simulate.}

\item{...}{other parameters passed to \code{\link{simulatePredictionData}}.}
}
\value{
a data frame with \code{nSamples} rows and columns with the
        predictive model accuracy.
}
\description{
Simulation for estimating logistic regression with uncertain responses.
}
