% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRatingMatrix.R
\name{simulateRatingMatrix2}
\alias{simulateRatingMatrix2}
\title{Simulate a rating matrix with two ratings per event.}
\usage{
simulateRatingMatrix2(
  nLevels,
  k,
  agree,
  nEvents = 100,
  response.probs = rep(1/nLevels, nLevels)
)
}
\arguments{
\item{nLevels}{the number of possible outcomes there are for each rating.}

\item{k}{the total number of available raters.}

\item{agree}{the percent of time the raters agree. Note that the actual
agreement of the simulated matrix will vary from this value
(see \link{sample}).}

\item{nEvents}{the number of rating events within each matrix.}

\item{response.probs}{probability weights for the distribution of scores. By
default, each of the levels has equal probability of being selected.
This allows situations where some responses are more common than
others (e.g. 50\% of students get a 3, 30\% get a 2, and 20\% get a 1).
This is independent of the percent agreement parameter.}
}
\value{
a data frame with a simulated rating matrix.
}
\description{
This is a alternative implementation of the [IRRsim::simulateRatingMatrix()\
function for when each event is has two ratings from k available raters.
For each event, one rater is randomly selected and given a rating from the
distribution defined by the \code{response.probs} parameter. A second randomly
selected rater will have the same rating with a probability defined by the
`agree` parameter.
}
