% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.IRRsim.R
\name{summary.IRRsim}
\alias{summary.IRRsim}
\title{Returns a summary of a IRR simulation.}
\usage{
\method{summary}{IRRsim}(object, method = "loess",
  agreements = seq(0.1, 0.9, by = 0.1), stat, k,
  predict.interval = "confidence", ...)
}
\arguments{
\item{method}{the method used to find predicted values. Possible values are
loess, linear, and quadratic.}

\item{agreements}{vector of percent agreements to include in the summary table.}

\item{stat}{the IRR statistic to return summary for, or missing for all metrics.}

\item{k}{which number of raters to print summary of.}

\item{...}{currently unused.}

\item{x}{the result of \code{\link{simulateICC}}.}
}
\description{
This function will provide predicted IRR statistics for various percent
agreements. There are two elements in the returned list: \code{model} and
\code{summary}. The former are the results of the modeling function and the
latter is a \code{data.frame} providing the predicted values for various
percent agreements. If \code{stat = 'all'} is specified, the summary
table contains the predicted values. If a specific IRR statistic is specified,
the summary table contains the predicted value along with the interval.#'
}
