% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.IRRsim.R
\name{as.data.frame.IRRsim}
\alias{as.data.frame.IRRsim}
\title{Converts the results of \code{\link{simulateICC}} to a \code{data.frame}}
\usage{
\method{as.data.frame}{IRRsim}(x, ...)
}
\arguments{
\item{x}{the result of \code{\link{simulateICC}}.}

\item{...}{currently unused.}
}
\description{
This function will create a data frame with all the IRR metrics combiend.
}
