% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateRatingMatrix.R
\name{simulateRatingMatrix}
\alias{simulateRatingMatrix}
\title{Simulate a rating matrix.}
\usage{
simulateRatingMatrix(
  nLevels,
  k,
  k_per_event = 2,
  agree,
  nEvents = 100,
  response.probs = rep(1/nLevels, nLevels)
)
}
\arguments{
\item{nLevels}{the number of possible outcomes there are for each rating.}

\item{k}{the total number of available raters.}

\item{k_per_event}{number of raters per scoring event.}

\item{agree}{the percent of time the raters agree. Note that the actual
agreement of the simulated matrix will vary from this value
(see \link{sample}).}

\item{nEvents}{the number of rating events within each matrix.}

\item{response.probs}{probability weights for the distribution of scores. By
default, each of the levels has equal probability of being selected.
This allows situations where some responses are more common than
others (e.g. 50\% of students get a 3, 30\% get a 2, and 20\% get a 1).
This is independent of the percent agreement parameter.}
}
\value{
a data frame with a simulated rating matrix.
}
\description{
This function will generate a \code{nEvents} x \code{k} scoring matrix.
}
\examples{
test <- simulateRatingMatrix(nLevels = 3, k = 2, agree = 0.6, nEvents = 100)
psych::ICC(test)
}
